requirements.txt
# Core dependencies
neo4j==5.15.0
networkx==3.2.1
pandas==2.1.4
numpy==1.26.2

# Machine Learning
torch==2.1.2
torch-geometric==2.5.0
scikit-learn==1.3.2

# Graph Embeddings
node2vec==0.4.6

# LLM and Agentic Pipeline
langchain==0.1.0
langchain-openai==0.0.5
openai==1.10.0

# Neo4j Python Driver
py2neo==2021.2.4

# Utilities
requests==2.31.0
matplotlib==3.8.2
tqdm==4.66.1
python-dotenv==1.0.0

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
notebook==7.0.6

